+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/User Input/LTR/ProgressBar/VisualToastNotificationComboBoxUserInputWithProgressBarForm.cs
@@ -81,9 +81,8 @@ internal partial class VisualToastNotificationComboBoxUserInputWithProgressBarFo
 
     private void UpdateLocation()
     {
-        //Once loaded, position the form, or position it to the bottom left of the screen with added padding
-        Location = _data.NotificationLocation ?? new Point(Screen.PrimaryScreen!.WorkingArea.Width - Width - 5,
-            Screen.PrimaryScreen.WorkingArea.Height - Height - 5);
+        // Once loaded, position the form, or default to bottom-right with DPI-scaled edge padding.
+        Location = _data.NotificationLocation ?? GetDefaultBottomRightLocation();
     }
 
     private void UpdateIcon()
@@ -97,14 +96,8 @@ internal partial class VisualToastNotificationComboBoxUserInputWithProgressBarFo
         SetIcon(bitmap);
     }
 
-    private void ShowCloseButton()
-    {
-        CloseBox = _data.ShowCloseBox ?? false;
-
-        FormBorderStyle = CloseBox ? FormBorderStyle.Fixed3D : FormBorderStyle.FixedSingle;
-
-        ControlBox = _data.ShowCloseBox ?? false;
-    }
+    private void ShowCloseButton() =>
+        ApplyCloseBoxChrome(_data.ShowCloseBox ?? false);
 
     private void VisualToastNotificationComboBoxUserInputWithProgressBarForm_Resize(object? sender, EventArgs e)
     {
@@ -118,9 +111,11 @@ internal partial class VisualToastNotificationComboBoxUserInputWithProgressBarFo
     {
         UpdateIcon();
 
-        UpdateLocation();
-
         ShowCloseButton();
+
+        ApplyToastDpiLayout();
+
+        UpdateLocation();
 
         kbtnDismiss.Text = KryptonManager.Strings.ToastNotificationStrings.Dismiss;